refactor(ingredient): extract whole-line recognizers into recognize.rs
Pure move, no behavior change. Lifts the three try_parse_* special-form
recognizers (optional-wrapped, trailing-amount, x-of-construction),
lift_inline_descriptive_paren, and is_temperature_unit out of pipeline.rs
into recognize.rs, along with the lift test. parse_core_ingredient is now
pub(super) so the recognizers can reach it. Second step of the split.
