+++ ingredient-parser/src/parser/mod.rs
@@ -6,6 +6,7 @@ pub(crate) mod helpers;
 pub(crate) mod measurement;
 pub(crate) mod normalize;
 pub(crate) mod pipeline;
+pub(crate) mod recognize;
 
 pub(crate) use helpers::parse_amount_string;
 pub(crate) use helpers::{
